+++ e2e/bash/controller.bash
@@ -22,6 +22,9 @@ teardown() {
 }
 
 @test "canister creation sets controller to the wallet" {
+    # invokes:
+    #  - wallet_create_canister
+
     assert_command dfx identity new alice
     assert_command dfx identity use alice
 
@@ -42,6 +45,9 @@ teardown() {
 }
 
 @test "canister installation sets controller to the wallet" {
+    # invokes:
+    #  - wallet_create_canister
+
     assert_command dfx identity new alice
     assert_command dfx identity use alice
 
@@ -61,6 +67,9 @@ teardown() {
 }
 
 @test "update-settings sets controller" {
+    # invokes:
+    #  - wallet_create_canister
+
     assert_command dfx identity new alice
     assert_command dfx identity new bob
 
@@ -91,12 +100,71 @@ teardown() {
     assert_command dfx --identity bob canister install e2e_project -m reinstall
 }
 
-@test "wallet create wallet" {
+@test "create wallet with single controller through wallet_create_wallet" {
+    # invokes:
+    #  - wallet_create_wallet
+    #  - update_settings_call
+    #  - update_settings_call (with controller)
+
     # curious: the cycles wallet has a wallet_create_wallet method, published
     # in its .did file.  dfx doesn't call it.  Maybe other users of the agent call it, though.
-    # The sdk repo has one call to the method in an e2e ref test.  This is a copy of that test.
+    # The sdk repo has one call to the method in an e2e ref test.  This is a copy of that test,
+    # there called "wallet create wallet".
     WALLET_ID=$(dfx identity get-wallet)
     CREATE_RES=$(dfx canister --no-wallet call "${WALLET_ID}" wallet_create_wallet "(record { cycles = (2000000000000:nat64); settings = record {controller = opt principal \"$(dfx identity get-principal)\";};})")
     CHILD_ID=$(echo "${CREATE_RES}" | tr '\n' ' ' |  cut -d'"' -f 2)
     assert_command dfx canister --no-wallet call "${CHILD_ID}" wallet_balance '()'
 }
+
+@test "create wallet with multiple controllers through wallet_create_wallet" {
+    # invokes:
+    #  - wallet_create_wallet
+    #  - update_settings_call
+    #  - update_settings_call (with controllers)
+
+    assert_command dfx identity new alice
+    assert_command dfx identity new bob
+
+    WALLET_ID=$(dfx identity get-wallet)
+    CREATE_RES=$(dfx canister --no-wallet call "${WALLET_ID}" wallet_create_wallet "(record { cycles = (2000000000000:nat64); settings = record {controllers = opt vec { principal \"$(dfx identity get-principal)\"; principal \"$(dfx --identity alice identity get-principal)\";};};})")
+    CHILD_ID=$(echo "${CREATE_RES}" | tr '\n' ' ' |  cut -d'"' -f 2)
+
+    assert_command dfx canister --no-wallet call "${CHILD_ID}" wallet_balance '()'
+    assert_command dfx --identity alice canister --no-wallet call "${CHILD_ID}" wallet_balance '()'
+    assert_command_fail dfx --identity bob canister --no-wallet call "${CHILD_ID}" wallet_balance '()'
+
+    assert_command dfx canister --no-wallet call "${CHILD_ID}" get_custodians '()'
+    assert_eq '(vec {})'
+    assert_command dfx canister --no-wallet call "${CHILD_ID}" get_controllers '()'
+    assert_match 'principal "'"$(dfx identity get-principal)"'"';
+    assert_match 'principal "'"$(dfx --identity alice identity get-principal)"'"';
+}
+
+@test "create wallet with multiple controllers, other than caller, through wallet_create_wallet" {
+    # invokes:
+    #  - wallet_create_wallet
+    #  - update_settings_call
+    #  - update_settings_call (with controllers)
+
+    assert_command dfx identity new alice
+    assert_command dfx identity new bob
+
+    WALLET_ID=$(dfx identity get-wallet)
+    CREATE_RES=$(dfx canister --no-wallet call "${WALLET_ID}" wallet_create_wallet "(record { cycles = (2000000000000:nat64); settings = record {controllers = opt vec { principal \"$(dfx --identity alice identity get-principal)\"; principal \"$(dfx --identity bob identity get-principal)\";};};})")
+    CHILD_ID=$(echo "${CREATE_RES}" | tr '\n' ' ' |  cut -d'"' -f 2)
+
+    assert_command_fail dfx canister --no-wallet call "${CHILD_ID}" wallet_balance '()'
+    assert_command dfx --identity alice canister --no-wallet call "${CHILD_ID}" wallet_balance '()'
+    assert_command dfx --identity bob canister --no-wallet call "${CHILD_ID}" wallet_balance '()'
+
+    assert_command_fail dfx canister --no-wallet call "${CHILD_ID}" get_custodians '()'
+    assert_command dfx --identity bob canister --no-wallet call "${CHILD_ID}" get_custodians '()'
+    assert_eq '(vec {})'
+
+    assert_command_fail dfx canister --no-wallet call "${CHILD_ID}" get_controllers '()'
+    assert_command dfx --identity alice canister --no-wallet call "${CHILD_ID}" get_controllers '()'
+    assert_not_match "$(dfx identity get-principal)"
+    assert_match 'principal "'"$(dfx --identity bob identity get-principal)"'"';
+    assert_match 'principal "'"$(dfx --identity alice identity get-principal)"'"';
+}
+